Split out network system calls in the libc backend.
